+++ Packages/com.unity.cloud.gltfast.tests/Tests/Runtime/Scripts/Import/ImportAddonInstanceCollectionTests.cs
@@ -2,18 +2,30 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using System;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
 using GLTFast.Addons;
+using GLTFast.Schema;
 using NUnit.Framework;
+using Unity.Collections;
+using Unity.PerformanceTesting;
 using UnityEngine;
 using UnityEngine.Profiling;
+using Texture = GLTFast.Schema.Texture;
 
 namespace GLTFast.Tests
 {
     class ImportAddonInstanceCollectionTests
     {
+        const int k_TextureCount = 100;
+        const int k_AddonCount = 10;
         const string k_ExtensionName = "ext";
 
         ImportAddonInstanceCollection m_Collection;
+        ImportAddonInstanceCollection m_TextureAddons;
+
+        Texture[] m_Textures = new Texture[k_TextureCount];
 
         AddonA m_AddonA;
         AddonB m_AddonB;
@@ -31,6 +43,21 @@ namespace GLTFast.Tests
             m_Collection.Add(m_AddonA);
             m_Collection.Add(m_AddonB);
             m_Collection.Add(m_AddonC);
+
+            m_Textures = new Texture[k_TextureCount];
+            for (var i = 0; i < k_TextureCount; i++)
+            {
+                m_Textures[i] = new Texture { source = i };
+            }
+
+            m_TextureAddons = new ImportAddonInstanceCollection();
+            for (var i = 0; i < k_AddonCount; i++)
+            {
+                ImportAddonInstance addon = i >= k_AddonCount - 1
+                    ? new TextureAddon()
+                    : new AddonA();
+                m_TextureAddons.Add(addon);
+            }
         }
 
         [OneTimeTearDown]
@@ -47,6 +74,18 @@ namespace GLTFast.Tests
 
             var a = m_Collection.Get<AddonBase>();
             Assert.AreSame(m_AddonA, a);
+        }
+
+        static readonly Func<ImportAddonInstance, bool> s_SupportsGltfExtensionPredicate =
+            addon => addon.SupportsGltfExtension(k_ExtensionName);
+
+        [Test]
+        public void First()
+        {
+            Profiler.BeginSample("ImportAddonInstanceCollectionTests.Exists");
+            var addon = m_Collection.First(s_SupportsGltfExtensionPredicate);
+            Profiler.EndSample();
+            Assert.NotNull(addon);
         }
 
         [Test]
@@ -69,11 +108,68 @@ namespace GLTFast.Tests
             });
             Assert.AreEqual(3, count);
         }
+
+        [Test, Performance]
+        public void ForEachTryGet()
+        {
+            var sg = new SampleGroup("Time", SampleUnit.Microsecond);
+            Measure.Method(() =>
+                {
+                    bool OverridesImage(ITextureImageLoader loader, TextureBase texture, out int imageIndex)
+                    {
+                        return loader.IsAbleToLoad(texture, out imageIndex);
+                    }
+
+                    Dictionary<int, int> overrides = null;
+
+                    m_TextureAddons.ForEachTryGet<ITextureImageLoader, TextureBase, int>(
+                        m_Textures,
+                        OverridesImage,
+                        (addon, textureIndex, imageIndex) =>
+                        {
+                            overrides ??= new Dictionary<int, int>();
+                            overrides[textureIndex] = imageIndex;
+                        }
+                    );
+
+                    Assert.IsNotNull(overrides);
+                    Assert.AreEqual(k_TextureCount / 4, overrides.Count);
+                }
+                    )
+                .SampleGroup(sg)
+                .MeasurementCount(1000)
+                .Run();
+        }
     }
 
     class AddonA : AddonBase { }
     class AddonB : AddonBase { }
     class AddonC : AddonBase { }
+
+    class TextureAddon : AddonBase, ITextureImageLoader
+    {
+        public bool IsAbleToLoad(TextureBase texture, out int imageIndex)
+        {
+            if (texture.source % 4 == 0)
+            {
+                imageIndex = texture.source * 2;
+                return true;
+            }
+            imageIndex = -1;
+            return false;
+        }
+
+        public Task<ImageResult> LoadImage(
+            NativeArray<byte>.ReadOnly data,
+            bool linear,
+            bool readable,
+            bool generateMipMaps,
+            CancellationToken cancellationToken
+        )
+        {
+            throw new NotImplementedException();
+        }
+    }
 
     class AddonBase : ImportAddonInstance
     {